requirements.txt
asgiref==3.7.2
Django==4.2.6
django-dynamic-fixture==4.0.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
Faker==19.12.0
freezegun==1.2.2
inflection==0.5.1
iniconfig==2.0.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
